pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "nba-playoff-model"
version = "0.1.0"
description = "Monte Carlo bracket simulation model for NBA playoff outcomes"
requires-python = ">=3.9"
dependencies = [
    "pandas>=2.1",
    "pyarrow>=14.0",
    "numpy>=1.26",
    "scipy>=1.11",
    "scikit-learn>=1.3",
    "statsmodels>=0.14",
    "pyyaml>=6.0",
    "nba_api>=1.4",
    "requests>=2.31",
    "beautifulsoup4>=4.12",
    "playwright>=1.40",
    "tqdm>=4.66",
]

[project.optional-dependencies]
dashboard = ["streamlit>=1.29"]
dev = [
    "pytest>=7.4",
    "pytest-cov>=4.1",
    "ruff>=0.1",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.ruff]
line-length = 100
target-version = "py39"

[tool.ruff.lint]
select = ["E", "F", "W", "I"]
ignore = []

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "--tb=short"
